doc_parsing/layoutlm/requirements.txt
transformers
torch
PyMuPDF
Pillow
pydantic
python-dotenv
openai
numpy
